@@ -2,13 +2,15 @@ name: camera
 description: A Flutter plugin for getting information about and controlling the
   camera on Android and iOS. Supports previewing the camera feed, capturing images, capturing video,
   and streaming image buffers to dart.
-version: 0.6.3+4
+version: 0.7.0+4
 homepage: https://github.com/flutter/plugins/tree/master/packages/camera/camera
 
 dependencies:
   flutter:
     sdk: flutter
-  camera_platform_interface: ">=1.0.4 <1.1.0"
+  camera_platform_interface: ^1.5.0
+  pedantic: ^1.8.0
+  quiver: ^2.1.5
 
 dev_dependencies:
   path_provider: ^0.5.0
@@ -17,7 +19,6 @@ dev_dependencies:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  pedantic: ^1.8.0
   mockito: ^4.1.3
   plugin_platform_interface: ^1.0.3
 
@@ -31,5 +32,5 @@ flutter:
         pluginClass: CameraPlugin
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
