@@ -30,7 +30,8 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
 
   GoogleMapController controller;
   Map<PolygonId, Polygon> polygons = <PolygonId, Polygon>{};
-  int _polygonIdCounter = 1;
+  Map<PolygonId, double> polygonOffsets = <PolygonId, double>{};
+  int _polygonIdCounter = 0;
   PolygonId selectedPolygon;
 
   // Values when toggling polygon color
@@ -79,7 +80,6 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
     }
 
     final String polygonIdVal = 'polygon_id_$_polygonIdCounter';
-    _polygonIdCounter++;
     final PolygonId polygonId = PolygonId(polygonIdVal);
 
     final Polygon polygon = Polygon(
@@ -96,6 +96,9 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
 
     setState(() {
       polygons[polygonId] = polygon;
+      polygonOffsets[polygonId] = _polygonIdCounter.ceilToDouble();
+      // increment _polygonIdCounter to have unique polygon id each time
+      _polygonIdCounter++;
     });
   }
 
@@ -144,6 +147,22 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
     });
   }
 
+  void _addHoles() {
+    final Polygon polygon = polygons[selectedPolygon];
+    setState(() {
+      polygons[selectedPolygon] = polygon.copyWith(holesParam: _createHoles());
+    });
+  }
+
+  void _removeHoles() {
+    final Polygon polygon = polygons[selectedPolygon];
+    setState(() {
+      polygons[selectedPolygon] = polygon.copyWith(
+        holesParam: <List<LatLng>>[],
+      );
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return Column(
@@ -173,20 +192,20 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
                   children: <Widget>[
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        TextButton(
                           child: const Text('add'),
                           onPressed: _add,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('remove'),
                           onPressed: (selectedPolygon == null) ? null : _remove,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('toggle visible'),
                           onPressed:
                               (selectedPolygon == null) ? null : _toggleVisible,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('toggle geodesic'),
                           onPressed: (selectedPolygon == null)
                               ? null
@@ -196,18 +215,34 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
                     ),
                     Column(
                       children: <Widget>[
-                        FlatButton(
+                        TextButton(
+                          child: const Text('add holes'),
+                          onPressed: (selectedPolygon == null)
+                              ? null
+                              : ((polygons[selectedPolygon].holes.isNotEmpty)
+                                  ? null
+                                  : _addHoles),
+                        ),
+                        TextButton(
+                          child: const Text('remove holes'),
+                          onPressed: (selectedPolygon == null)
+                              ? null
+                              : ((polygons[selectedPolygon].holes.isEmpty)
+                                  ? null
+                                  : _removeHoles),
+                        ),
+                        TextButton(
                           child: const Text('change stroke width'),
                           onPressed:
                               (selectedPolygon == null) ? null : _changeWidth,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change stroke color'),
                           onPressed: (selectedPolygon == null)
                               ? null
                               : _changeStrokeColor,
                         ),
-                        FlatButton(
+                        TextButton(
                           child: const Text('change fill color'),
                           onPressed: (selectedPolygon == null)
                               ? null
@@ -233,6 +268,27 @@ class PlacePolygonBodyState extends State<PlacePolygonBody> {
     points.add(_createLatLng(52.4351 + offset, -4.5235));
     points.add(_createLatLng(52.1231 + offset, -5.0829));
     return points;
+  }
+
+  List<List<LatLng>> _createHoles() {
+    final List<List<LatLng>> holes = <List<LatLng>>[];
+    final double offset = polygonOffsets[selectedPolygon];
+
+    final List<LatLng> hole1 = <LatLng>[];
+    hole1.add(_createLatLng(51.8395 + offset, -3.8814));
+    hole1.add(_createLatLng(52.0234 + offset, -3.9914));
+    hole1.add(_createLatLng(52.1351 + offset, -4.4435));
+    hole1.add(_createLatLng(52.0231 + offset, -4.5829));
+    holes.add(hole1);
+
+    final List<LatLng> hole2 = <LatLng>[];
+    hole2.add(_createLatLng(52.2395 + offset, -3.6814));
+    hole2.add(_createLatLng(52.4234 + offset, -3.7914));
+    hole2.add(_createLatLng(52.5351 + offset, -4.2435));
+    hole2.add(_createLatLng(52.4231 + offset, -4.3829));
+    holes.add(hole2);
+
+    return holes;
   }
 
   LatLng _createLatLng(double lat, double lng) {
