@@ -1,6 +1,9 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.camera;
 
-import static android.view.OrientationEventListener.ORIENTATION_UNKNOWN;
 import static io.flutter.plugins.camera.CameraUtils.computeBestPreviewSize;
 
 import android.annotation.SuppressLint;
@@ -21,6 +24,7 @@ import android.hardware.camera2.CaptureFailure;
 import android.hardware.camera2.CaptureRequest;
 import android.hardware.camera2.CaptureResult;
 import android.hardware.camera2.TotalCaptureResult;
+import android.hardware.camera2.params.MeteringRectangle;
 import android.hardware.camera2.params.OutputConfiguration;
 import android.hardware.camera2.params.SessionConfiguration;
 import android.media.CamcorderProfile;
@@ -32,15 +36,21 @@ import android.os.Build.VERSION;
 import android.os.Build.VERSION_CODES;
 import android.os.Handler;
 import android.os.Looper;
+import android.util.Log;
+import android.util.Range;
+import android.util.Rational;
 import android.util.Size;
-import android.view.OrientationEventListener;
 import android.view.Surface;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugins.camera.PictureCaptureRequest.State;
 import io.flutter.plugins.camera.media.MediaRecorderBuilder;
+import io.flutter.plugins.camera.types.ExposureMode;
 import io.flutter.plugins.camera.types.FlashMode;
+import io.flutter.plugins.camera.types.FocusMode;
 import io.flutter.plugins.camera.types.ResolutionPreset;
 import io.flutter.view.TextureRegistry.SurfaceTextureEntry;
 import java.io.File;
@@ -55,10 +65,17 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.Executors;
 
+@FunctionalInterface
+interface ErrorCallback {
+  void onError(String errorCode, String errorMessage);
+}
+
 public class Camera {
+  private static final String TAG = "Camera";
+
   private final SurfaceTextureEntry flutterTexture;
   private final CameraManager cameraManager;
-  private final OrientationEventListener orientationEventListener;
+  private final DeviceOrientationManager deviceOrientationListener;
   private final boolean isFrontFacing;
   private final int sensorOrientation;
   private final String cameraName;
@@ -69,6 +86,7 @@ public class Camera {
   private final CamcorderProfile recordingProfile;
   private final DartMessenger dartMessenger;
   private final CameraZoom cameraZoom;
+  private final CameraCharacteristics cameraCharacteristics;
 
   private CameraDevice cameraDevice;
   private CameraCaptureSession cameraCaptureSession;
@@ -78,9 +96,23 @@ public class Camera {
   private MediaRecorder mediaRecorder;
   private boolean recordingVideo;
   private File videoRecordingFile;
-  private int currentOrientation = ORIENTATION_UNKNOWN;
   private FlashMode flashMode;
+  private ExposureMode exposureMode;
+  private FocusMode focusMode;
   private PictureCaptureRequest pictureCaptureRequest;
+  private CameraRegions cameraRegions;
+  private int exposureOffset;
+  private boolean useAutoFocus = true;
+  private Range<Integer> fpsRange;
+  private PlatformChannel.DeviceOrientation lockedCaptureOrientation;
+
+  private static final HashMap<String, Integer> supportedImageFormats;
+  // Current supported outputs
+  static {
+    supportedImageFormats = new HashMap<>();
+    supportedImageFormats.put("yuv420", 35);
+    supportedImageFormats.put("jpeg", 256);
+  }
 
   public Camera(
       final Activity activity,
@@ -100,23 +132,16 @@ public class Camera {
     this.cameraManager = (CameraManager) activity.getSystemService(Context.CAMERA_SERVICE);
     this.applicationContext = activity.getApplicationContext();
     this.flashMode = FlashMode.auto;
-    orientationEventListener =
-        new OrientationEventListener(activity.getApplicationContext()) {
-          @Override
-          public void onOrientationChanged(int i) {
-            if (i == ORIENTATION_UNKNOWN) {
-              return;
-            }
-            // Convert the raw deg angle to the nearest multiple of 90.
-            currentOrientation = (int) Math.round(i / 90.0) * 90;
-          }
-        };
-    orientationEventListener.enable();
+    this.exposureMode = ExposureMode.auto;
+    this.focusMode = FocusMode.auto;
+    this.exposureOffset = 0;
 
-    CameraCharacteristics characteristics = cameraManager.getCameraCharacteristics(cameraName);
-    sensorOrientation = characteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
+    cameraCharacteristics = cameraManager.getCameraCharacteristics(cameraName);
+    initFps(cameraCharacteristics);
+    sensorOrientation = cameraCharacteristics.get(CameraCharacteristics.SENSOR_ORIENTATION);
     isFrontFacing =
-        characteristics.get(CameraCharacteristics.LENS_FACING) == CameraMetadata.LENS_FACING_FRONT;
+        cameraCharacteristics.get(CameraCharacteristics.LENS_FACING)
+            == CameraMetadata.LENS_FACING_FRONT;
     ResolutionPreset preset = ResolutionPreset.valueOf(resolutionPreset);
     recordingProfile =
         CameraUtils.getBestAvailableCamcorderProfileForResolutionPreset(cameraName, preset);
@@ -124,8 +149,33 @@ public class Camera {
     previewSize = computeBestPreviewSize(cameraName, preset);
     cameraZoom =
         new CameraZoom(
-            characteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE),
-            characteristics.get(CameraCharacteristics.SCALER_AVAILABLE_MAX_DIGITAL_ZOOM));
+            cameraCharacteristics.get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE),
+            cameraCharacteristics.get(CameraCharacteristics.SCALER_AVAILABLE_MAX_DIGITAL_ZOOM));
+
+    deviceOrientationListener =
+        new DeviceOrientationManager(activity, dartMessenger, isFrontFacing, sensorOrientation);
+    deviceOrientationListener.start();
+  }
+
+  private void initFps(CameraCharacteristics cameraCharacteristics) {
+    try {
+      Range<Integer>[] ranges =
+          cameraCharacteristics.get(CameraCharacteristics.CONTROL_AE_AVAILABLE_TARGET_FPS_RANGES);
+      if (ranges != null) {
+        for (Range<Integer> range : ranges) {
+          int upper = range.getUpper();
+          Log.i("Camera", "[FPS Range Available] is:" + range);
+          if (upper >= 10) {
+            if (fpsRange == null || upper > fpsRange.getUpper()) {
+              fpsRange = range;
+            }
+          }
+        }
+      }
+    } catch (Exception e) {
+      e.printStackTrace();
+    }
+    Log.i("Camera", "[FPS Range] is:" + fpsRange);
   }
 
   private void prepareMediaRecorder(String outputFilePath) throws IOException {
@@ -136,20 +186,28 @@ public class Camera {
     mediaRecorder =
         new MediaRecorderBuilder(recordingProfile, outputFilePath)
             .setEnableAudio(enableAudio)
-            .setMediaOrientation(getMediaOrientation())
+            .setMediaOrientation(
+                lockedCaptureOrientation == null
+                    ? deviceOrientationListener.getMediaOrientation()
+                    : deviceOrientationListener.getMediaOrientation(lockedCaptureOrientation))
             .build();
   }
 
   @SuppressLint("MissingPermission")
-  public void open() throws CameraAccessException {
+  public void open(String imageFormatGroup) throws CameraAccessException {
     pictureImageReader =
         ImageReader.newInstance(
             captureSize.getWidth(), captureSize.getHeight(), ImageFormat.JPEG, 2);
 
+    Integer imageFormat = supportedImageFormats.get(imageFormatGroup);
+    if (imageFormat == null) {
+      Log.w(TAG, "The selected imageFormatGroup is not supported by Android. Defaulting to yuv420");
+      imageFormat = ImageFormat.YUV_420_888;
+    }
+
     // Used to steam image byte data to dart side.
     imageStreamReader =
-        ImageReader.newInstance(
-            previewSize.getWidth(), previewSize.getHeight(), ImageFormat.YUV_420_888, 2);
+        ImageReader.newInstance(previewSize.getWidth(), previewSize.getHeight(), imageFormat, 2);
 
     cameraManager.openCamera(
         cameraName,
@@ -159,14 +217,17 @@ public class Camera {
             cameraDevice = device;
             try {
               startPreview();
+              dartMessenger.sendCameraInitializedEvent(
+                  previewSize.getWidth(),
+                  previewSize.getHeight(),
+                  exposureMode,
+                  focusMode,
+                  isExposurePointSupported(),
+                  isFocusPointSupported());
             } catch (CameraAccessException e) {
               dartMessenger.sendCameraErrorEvent(e.getMessage());
               close();
-              return;
             }
-
-            dartMessenger.sendCameraInitializedEvent(
-                previewSize.getWidth(), previewSize.getHeight());
           }
 
           @Override
@@ -210,219 +271,6 @@ public class Camera {
         null);
   }
 
-  private void writeToFile(ByteBuffer buffer, File file) throws IOException {
-    try (FileOutputStream outputStream = new FileOutputStream(file)) {
-      while (0 < buffer.remaining()) {
-        outputStream.getChannel().write(buffer);
-      }
-    }
-  }
-
-  public void takePicture(@NonNull final Result result) {
-    // Only take 1 picture at a time
-    if (pictureCaptureRequest != null && !pictureCaptureRequest.isFinished()) {
-      result.error("captureAlreadyActive", "Picture is currently already being captured", null);
-      return;
-    }
-    // Store the result
-    this.pictureCaptureRequest = new PictureCaptureRequest(result);
-
-    // Create temporary file
-    final File outputDir = applicationContext.getCacheDir();
-    final File file;
-    try {
-      file = File.createTempFile("CAP", ".jpg", outputDir);
-    } catch (IOException | SecurityException e) {
-      pictureCaptureRequest.error("cannotCreateFile", e.getMessage(), null);
-      return;
-    }
-
-    // Listen for picture being taken
-    pictureImageReader.setOnImageAvailableListener(
-        reader -> {
-          try (Image image = reader.acquireLatestImage()) {
-            ByteBuffer buffer = image.getPlanes()[0].getBuffer();
-            writeToFile(buffer, file);
-            pictureCaptureRequest.finish(file.getAbsolutePath());
-          } catch (IOException e) {
-            pictureCaptureRequest.error("IOError", "Failed saving image", null);
-          }
-        },
-        null);
-
-    runPictureAutoFocus();
-  }
-
-  private final CameraCaptureSession.CaptureCallback pictureCaptureCallback =
-      new CameraCaptureSession.CaptureCallback() {
-        @Override
-        public void onCaptureCompleted(
-            @NonNull CameraCaptureSession session,
-            @NonNull CaptureRequest request,
-            @NonNull TotalCaptureResult result) {
-          processCapture(result);
-        }
-
-        @Override
-        public void onCaptureProgressed(
-            @NonNull CameraCaptureSession session,
-            @NonNull CaptureRequest request,
-            @NonNull CaptureResult partialResult) {
-          processCapture(partialResult);
-        }
-
-        @Override
-        public void onCaptureFailed(
-            @NonNull CameraCaptureSession session,
-            @NonNull CaptureRequest request,
-            @NonNull CaptureFailure failure) {
-          if (pictureCaptureRequest == null || pictureCaptureRequest.isFinished()) {
-            return;
-          }
-          String reason;
-          switch (failure.getReason()) {
-            case CaptureFailure.REASON_ERROR:
-              reason = "An error happened in the framework";
-              break;
-            case CaptureFailure.REASON_FLUSHED:
-              reason = "The capture has failed due to an abortCaptures() call";
-              break;
-            default:
-              reason = "Unknown reason";
-          }
-          pictureCaptureRequest.error("captureFailure", reason, null);
-        }
-
-        private void processCapture(CaptureResult result) {
-          if (pictureCaptureRequest == null) {
-            return;
-          }
-
-          Integer aeState = result.get(CaptureResult.CONTROL_AE_STATE);
-          Integer afState = result.get(CaptureResult.CONTROL_AF_STATE);
-          switch (pictureCaptureRequest.getState()) {
-            case focusing:
-              if (afState == null) {
-                return;
-              } else if (afState == CaptureResult.CONTROL_AF_STATE_FOCUSED_LOCKED
-                  || afState == CaptureResult.CONTROL_AF_STATE_NOT_FOCUSED_LOCKED) {
-                // Some devices might return null here, in which case we will also continue.
-                if (aeState == null || aeState == CaptureResult.CONTROL_AE_STATE_CONVERGED) {
-                  runPictureCapture();
-                } else {
-                  runPicturePreCapture();
-                }
-              }
-              break;
-            case preCapture:
-              // Some devices might return null here, in which case we will also continue.
-              if (aeState == null
-                  || aeState == CaptureRequest.CONTROL_AE_STATE_PRECAPTURE
-                  || aeState == CaptureRequest.CONTROL_AE_STATE_FLASH_REQUIRED
-                  || aeState == CaptureRequest.CONTROL_AE_STATE_CONVERGED) {
-                pictureCaptureRequest.setState(State.waitingPreCaptureReady);
-              }
-              break;
-            case waitingPreCaptureReady:
-              if (aeState == null || aeState != CaptureRequest.CONTROL_AE_STATE_PRECAPTURE) {
-                runPictureCapture();
-              }
-          }
-        }
-      };
-
-  private void runPictureAutoFocus() {
-    assert (pictureCaptureRequest != null);
-    pictureCaptureRequest.setState(PictureCaptureRequest.State.focusing);
-    lockAutoFocus();
-  }
-
-  private void runPicturePreCapture() {
-    assert (pictureCaptureRequest != null);
-    pictureCaptureRequest.setState(PictureCaptureRequest.State.preCapture);
-
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
-        CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_START);
-    try {
-      cameraCaptureSession.capture(captureRequestBuilder.build(), pictureCaptureCallback, null);
-      captureRequestBuilder.set(
-          CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
-          CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_IDLE);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
-  }
-
-  private void runPictureCapture() {
-    assert (pictureCaptureRequest != null);
-    pictureCaptureRequest.setState(PictureCaptureRequest.State.capturing);
-    try {
-      final CaptureRequest.Builder captureBuilder =
-          cameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_STILL_CAPTURE);
-      captureBuilder.addTarget(pictureImageReader.getSurface());
-      captureBuilder.set(CaptureRequest.JPEG_ORIENTATION, getMediaOrientation());
-      switch (flashMode) {
-        case off:
-          captureBuilder.set(CaptureRequest.CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_ON);
-          captureBuilder.set(CaptureRequest.FLASH_MODE, CaptureRequest.FLASH_MODE_OFF);
-          break;
-        case auto:
-          captureBuilder.set(
-              CaptureRequest.CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_ON_AUTO_FLASH);
-          break;
-        case always:
-        default:
-          captureBuilder.set(
-              CaptureRequest.CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_ON_ALWAYS_FLASH);
-          break;
-      }
-      cameraCaptureSession.stopRepeating();
-      cameraCaptureSession.capture(
-          captureBuilder.build(),
-          new CameraCaptureSession.CaptureCallback() {
-            @Override
-            public void onCaptureCompleted(
-                @NonNull CameraCaptureSession session,
-                @NonNull CaptureRequest request,
-                @NonNull TotalCaptureResult result) {
-              unlockAutoFocus();
-            }
-          },
-          null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
-  }
-
-  private void lockAutoFocus() {
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_START);
-    try {
-      cameraCaptureSession.capture(captureRequestBuilder.build(), pictureCaptureCallback, null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
-  }
-
-  private void unlockAutoFocus() {
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_CANCEL);
-    initPreviewCaptureBuilder();
-    try {
-      cameraCaptureSession.capture(captureRequestBuilder.build(), null, null);
-    } catch (CameraAccessException ignored) {
-    }
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_IDLE);
-    try {
-      cameraCaptureSession.setRepeatingRequest(
-          captureRequestBuilder.build(), pictureCaptureCallback, null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
-  }
-
   private void createCaptureSession(int templateType, Surface... surfaces)
       throws CameraAccessException {
     createCaptureSession(templateType, null, surfaces);
@@ -451,28 +299,26 @@ public class Camera {
       }
     }
 
+    cameraRegions = new CameraRegions(getRegionBoundaries());
+
     // Prepare the callback
     CameraCaptureSession.StateCallback callback =
         new CameraCaptureSession.StateCallback() {
           @Override
           public void onConfigured(@NonNull CameraCaptureSession session) {
-            try {
-              if (cameraDevice == null) {
-                dartMessenger.sendCameraErrorEvent("The camera was closed during configuration.");
-                return;
-              }
-              cameraCaptureSession = session;
-              initPreviewCaptureBuilder();
-              cameraCaptureSession.setRepeatingRequest(
-                  captureRequestBuilder.build(),
-                  pictureCaptureCallback,
-                  new Handler(Looper.getMainLooper()));
-              if (onSuccessCallback != null) {
-                onSuccessCallback.run();
-              }
-            } catch (CameraAccessException | IllegalStateException | IllegalArgumentException e) {
-              dartMessenger.sendCameraErrorEvent(e.getMessage());
+            if (cameraDevice == null) {
+              dartMessenger.sendCameraErrorEvent("The camera was closed during configuration.");
+              return;
             }
+            cameraCaptureSession = session;
+
+            updateFpsRange();
+            updateFocus(focusMode);
+            updateFlash(flashMode);
+            updateExposure(exposureMode);
+
+            refreshPreviewCaptureSession(
+                onSuccessCallback, (code, message) -> dartMessenger.sendCameraErrorEvent(message));
           }
 
           @Override
@@ -519,6 +365,250 @@ public class Camera {
     cameraDevice.createCaptureSession(surfaces, callback, null);
   }
 
+  private void refreshPreviewCaptureSession(
+      @Nullable Runnable onSuccessCallback, @NonNull ErrorCallback onErrorCallback) {
+    if (cameraCaptureSession == null) {
+      return;
+    }
+
+    try {
+      cameraCaptureSession.setRepeatingRequest(
+          captureRequestBuilder.build(),
+          pictureCaptureCallback,
+          new Handler(Looper.getMainLooper()));
+
+      if (onSuccessCallback != null) {
+        onSuccessCallback.run();
+      }
+    } catch (CameraAccessException | IllegalStateException | IllegalArgumentException e) {
+      onErrorCallback.onError("cameraAccess", e.getMessage());
+    }
+  }
+
+  private void writeToFile(ByteBuffer buffer, File file) throws IOException {
+    try (FileOutputStream outputStream = new FileOutputStream(file)) {
+      while (0 < buffer.remaining()) {
+        outputStream.getChannel().write(buffer);
+      }
+    }
+  }
+
+  public void takePicture(@NonNull final Result result) {
+    // Only take 1 picture at a time
+    if (pictureCaptureRequest != null && !pictureCaptureRequest.isFinished()) {
+      result.error("captureAlreadyActive", "Picture is currently already being captured", null);
+      return;
+    }
+    // Store the result
+    this.pictureCaptureRequest = new PictureCaptureRequest(result);
+
+    // Create temporary file
+    final File outputDir = applicationContext.getCacheDir();
+    final File file;
+    try {
+      file = File.createTempFile("CAP", ".jpg", outputDir);
+    } catch (IOException | SecurityException e) {
+      pictureCaptureRequest.error("cannotCreateFile", e.getMessage(), null);
+      return;
+    }
+
+    // Listen for picture being taken
+    pictureImageReader.setOnImageAvailableListener(
+        reader -> {
+          try (Image image = reader.acquireLatestImage()) {
+            ByteBuffer buffer = image.getPlanes()[0].getBuffer();
+            writeToFile(buffer, file);
+            pictureCaptureRequest.finish(file.getAbsolutePath());
+          } catch (IOException e) {
+            pictureCaptureRequest.error("IOError", "Failed saving image", null);
+          }
+        },
+        null);
+
+    if (useAutoFocus) {
+      runPictureAutoFocus();
+    } else {
+      runPicturePreCapture();
+    }
+  }
+
+  private final CameraCaptureSession.CaptureCallback pictureCaptureCallback =
+      new CameraCaptureSession.CaptureCallback() {
+        @Override
+        public void onCaptureCompleted(
+            @NonNull CameraCaptureSession session,
+            @NonNull CaptureRequest request,
+            @NonNull TotalCaptureResult result) {
+          processCapture(result);
+        }
+
+        @Override
+        public void onCaptureProgressed(
+            @NonNull CameraCaptureSession session,
+            @NonNull CaptureRequest request,
+            @NonNull CaptureResult partialResult) {
+          processCapture(partialResult);
+        }
+
+        @Override
+        public void onCaptureFailed(
+            @NonNull CameraCaptureSession session,
+            @NonNull CaptureRequest request,
+            @NonNull CaptureFailure failure) {
+          if (pictureCaptureRequest == null || pictureCaptureRequest.isFinished()) {
+            return;
+          }
+          String reason;
+          boolean fatalFailure = false;
+          switch (failure.getReason()) {
+            case CaptureFailure.REASON_ERROR:
+              reason = "An error happened in the framework";
+              break;
+            case CaptureFailure.REASON_FLUSHED:
+              reason = "The capture has failed due to an abortCaptures() call";
+              fatalFailure = true;
+              break;
+            default:
+              reason = "Unknown reason";
+          }
+          Log.w("Camera", "pictureCaptureCallback.onCaptureFailed(): " + reason);
+          if (fatalFailure) pictureCaptureRequest.error("captureFailure", reason, null);
+        }
+
+        private void processCapture(CaptureResult result) {
+          if (pictureCaptureRequest == null) {
+            return;
+          }
+
+          Integer aeState = result.get(CaptureResult.CONTROL_AE_STATE);
+          Integer afState = result.get(CaptureResult.CONTROL_AF_STATE);
+          switch (pictureCaptureRequest.getState()) {
+            case focusing:
+              if (afState == null) {
+                return;
+              } else if (afState == CaptureResult.CONTROL_AF_STATE_FOCUSED_LOCKED
+                  || afState == CaptureResult.CONTROL_AF_STATE_NOT_FOCUSED_LOCKED) {
+                // Some devices might return null here, in which case we will also continue.
+                if (aeState == null || aeState == CaptureResult.CONTROL_AE_STATE_CONVERGED) {
+                  runPictureCapture();
+                } else {
+                  runPicturePreCapture();
+                }
+              }
+              break;
+            case preCapture:
+              // Some devices might return null here, in which case we will also continue.
+              if (aeState == null
+                  || aeState == CaptureRequest.CONTROL_AE_STATE_PRECAPTURE
+                  || aeState == CaptureRequest.CONTROL_AE_STATE_FLASH_REQUIRED
+                  || aeState == CaptureRequest.CONTROL_AE_STATE_CONVERGED) {
+                pictureCaptureRequest.setState(State.waitingPreCaptureReady);
+              }
+              break;
+            case waitingPreCaptureReady:
+              if (aeState == null || aeState != CaptureRequest.CONTROL_AE_STATE_PRECAPTURE) {
+                runPictureCapture();
+              }
+          }
+        }
+      };
+
+  private void runPictureAutoFocus() {
+    assert (pictureCaptureRequest != null);
+
+    pictureCaptureRequest.setState(PictureCaptureRequest.State.focusing);
+    lockAutoFocus(pictureCaptureCallback);
+  }
+
+  private void runPicturePreCapture() {
+    assert (pictureCaptureRequest != null);
+    pictureCaptureRequest.setState(PictureCaptureRequest.State.preCapture);
+
+    captureRequestBuilder.set(
+        CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
+        CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_START);
+
+    refreshPreviewCaptureSession(
+        () ->
+            captureRequestBuilder.set(
+                CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER,
+                CaptureRequest.CONTROL_AE_PRECAPTURE_TRIGGER_IDLE),
+        (code, message) -> pictureCaptureRequest.error(code, message, null));
+  }
+
+  private void runPictureCapture() {
+    assert (pictureCaptureRequest != null);
+    pictureCaptureRequest.setState(PictureCaptureRequest.State.capturing);
+    try {
+      final CaptureRequest.Builder captureBuilder =
+          cameraDevice.createCaptureRequest(CameraDevice.TEMPLATE_STILL_CAPTURE);
+      captureBuilder.addTarget(pictureImageReader.getSurface());
+      captureBuilder.set(
+          CaptureRequest.SCALER_CROP_REGION,
+          captureRequestBuilder.get(CaptureRequest.SCALER_CROP_REGION));
+      captureBuilder.set(
+          CaptureRequest.JPEG_ORIENTATION,
+          lockedCaptureOrientation == null
+              ? deviceOrientationListener.getMediaOrientation()
+              : deviceOrientationListener.getMediaOrientation(lockedCaptureOrientation));
+
+      switch (flashMode) {
+        case off:
+          captureBuilder.set(CaptureRequest.CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_ON);
+          captureBuilder.set(CaptureRequest.FLASH_MODE, CaptureRequest.FLASH_MODE_OFF);
+          break;
+        case auto:
+          captureBuilder.set(
+              CaptureRequest.CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_ON_AUTO_FLASH);
+          break;
+        case always:
+        default:
+          captureBuilder.set(
+              CaptureRequest.CONTROL_AE_MODE, CaptureRequest.CONTROL_AE_MODE_ON_ALWAYS_FLASH);
+          break;
+      }
+      cameraCaptureSession.stopRepeating();
+      cameraCaptureSession.capture(
+          captureBuilder.build(),
+          new CameraCaptureSession.CaptureCallback() {
+            @Override
+            public void onCaptureCompleted(
+                @NonNull CameraCaptureSession session,
+                @NonNull CaptureRequest request,
+                @NonNull TotalCaptureResult result) {
+              unlockAutoFocus();
+            }
+          },
+          null);
+    } catch (CameraAccessException e) {
+      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
+    }
+  }
+
+  private void lockAutoFocus(CaptureCallback callback) {
+    captureRequestBuilder.set(
+        CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_START);
+
+    refreshPreviewCaptureSession(
+        null, (code, message) -> pictureCaptureRequest.error(code, message, null));
+  }
+
+  private void unlockAutoFocus() {
+    captureRequestBuilder.set(
+        CaptureRequest.CONTROL_AF_TRIGGER, CameraMetadata.CONTROL_AF_TRIGGER_CANCEL);
+    updateFocus(focusMode);
+    try {
+      cameraCaptureSession.capture(captureRequestBuilder.build(), null, null);
+    } catch (CameraAccessException ignored) {
+    }
+    captureRequestBuilder.set(
+        CaptureRequest.CONTROL_AF_TRIGGER, CaptureRequest.CONTROL_AF_TRIGGER_IDLE);
+
+    refreshPreviewCaptureSession(
+        null,
+        (errorCode, errorMessage) -> pictureCaptureRequest.error(errorCode, errorMessage, null));
+  }
+
   public void startVideoRecording(Result result) {
     final File outputDir = applicationContext.getCacheDir();
     try {
@@ -549,7 +639,14 @@ public class Camera {
 
     try {
       recordingVideo = false;
-      mediaRecorder.stop();
+
+      try {
+        cameraCaptureSession.abortCaptures();
+        mediaRecorder.stop();
+      } catch (CameraAccessException | IllegalStateException e) {
+        // Ignore exceptions and try to continue (changes are camera session already aborted capture)
+      }
+
       mediaRecorder.reset();
       startPreview();
       result.success(videoRecordingFile.getAbsolutePath());
@@ -605,16 +702,11 @@ public class Camera {
   public void setFlashMode(@NonNull final Result result, FlashMode mode)
       throws CameraAccessException {
     // Get the flash availability
-    Boolean flashAvailable;
-    try {
-      flashAvailable =
-          cameraManager
-              .getCameraCharacteristics(cameraDevice.getId())
-              .get(CameraCharacteristics.FLASH_INFO_AVAILABLE);
-    } catch (CameraAccessException e) {
-      result.error("setFlashModeFailed", e.getMessage(), null);
-      return;
-    }
+    Boolean flashAvailable =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.FLASH_INFO_AVAILABLE);
+
     // Check if flash is available.
     if (flashAvailable == null || !flashAvailable) {
       result.error("setFlashModeFailed", "Device does not have flash capabilities", null);
@@ -623,8 +715,8 @@ public class Camera {
 
     // If switching directly from torch to auto or on, make sure we turn off the torch.
     if (flashMode == FlashMode.torch && mode != FlashMode.torch && mode != FlashMode.off) {
-      this.flashMode = FlashMode.off;
-      initPreviewCaptureBuilder();
+      updateFlash(FlashMode.off);
+
       this.cameraCaptureSession.setRepeatingRequest(
           captureRequestBuilder.build(),
           new CaptureCallback() {
@@ -640,8 +732,13 @@ public class Camera {
               }
 
               updateFlash(mode);
-              result.success(null);
-              isFinished = true;
+              refreshPreviewCaptureSession(
+                  () -> {
+                    result.success(null);
+                    isFinished = true;
+                  },
+                  (code, message) ->
+                      result.error("setFlashModeFailed", "Could not set flash mode.", null));
             }
 
             @Override
@@ -660,24 +757,302 @@ public class Camera {
           null);
     } else {
       updateFlash(mode);
-      result.success(null);
+
+      refreshPreviewCaptureSession(
+          () -> result.success(null),
+          (code, message) -> result.error("setFlashModeFailed", "Could not set flash mode.", null));
     }
+  }
+
+  public void setExposureMode(@NonNull final Result result, ExposureMode mode)
+      throws CameraAccessException {
+    updateExposure(mode);
+    cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
+    result.success(null);
+  }
+
+  public void setExposurePoint(@NonNull final Result result, Double x, Double y)
+      throws CameraAccessException {
+    // Check if exposure point functionality is available.
+    if (!isExposurePointSupported()) {
+      result.error(
+          "setExposurePointFailed", "Device does not have exposure point capabilities", null);
+      return;
+    }
+    // Check if the current region boundaries are known
+    if (cameraRegions.getMaxBoundaries() == null) {
+      result.error("setExposurePointFailed", "Could not determine max region boundaries", null);
+      return;
+    }
+    // Set the metering rectangle
+    if (x == null || y == null) cameraRegions.resetAutoExposureMeteringRectangle();
+    else cameraRegions.setAutoExposureMeteringRectangleFromPoint(x, y);
+    // Apply it
+    updateExposure(exposureMode);
+    refreshPreviewCaptureSession(
+        () -> result.success(null), (code, message) -> result.error("CameraAccess", message, null));
+  }
+
+  public void setFocusMode(@NonNull final Result result, FocusMode mode)
+      throws CameraAccessException {
+    this.focusMode = mode;
+
+    updateFocus(mode);
+
+    switch (mode) {
+      case auto:
+        refreshPreviewCaptureSession(
+            null, (code, message) -> result.error("setFocusMode", message, null));
+        break;
+      case locked:
+        lockAutoFocus(
+            new CaptureCallback() {
+              @Override
+              public void onCaptureCompleted(
+                  @NonNull CameraCaptureSession session,
+                  @NonNull CaptureRequest request,
+                  @NonNull TotalCaptureResult result) {
+                unlockAutoFocus();
+              }
+            });
+        break;
+    }
+    result.success(null);
+  }
+
+  public void setFocusPoint(@NonNull final Result result, Double x, Double y)
+      throws CameraAccessException {
+    // Check if focus point functionality is available.
+    if (!isFocusPointSupported()) {
+      result.error("setFocusPointFailed", "Device does not have focus point capabilities", null);
+      return;
+    }
+
+    // Check if the current region boundaries are known
+    if (cameraRegions.getMaxBoundaries() == null) {
+      result.error("setFocusPointFailed", "Could not determine max region boundaries", null);
+      return;
+    }
+
+    // Set the metering rectangle
+    if (x == null || y == null) {
+      cameraRegions.resetAutoFocusMeteringRectangle();
+    } else {
+      cameraRegions.setAutoFocusMeteringRectangleFromPoint(x, y);
+    }
+
+    // Apply the new metering rectangle
+    setFocusMode(result, focusMode);
+  }
+
+  @TargetApi(VERSION_CODES.P)
+  private boolean supportsDistortionCorrection() throws CameraAccessException {
+    int[] availableDistortionCorrectionModes =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.DISTORTION_CORRECTION_AVAILABLE_MODES);
+    if (availableDistortionCorrectionModes == null) availableDistortionCorrectionModes = new int[0];
+    long nonOffModesSupported =
+        Arrays.stream(availableDistortionCorrectionModes)
+            .filter((value) -> value != CaptureRequest.DISTORTION_CORRECTION_MODE_OFF)
+            .count();
+    return nonOffModesSupported > 0;
+  }
+
+  private Size getRegionBoundaries() throws CameraAccessException {
+    // No distortion correction support
+    if (android.os.Build.VERSION.SDK_INT < VERSION_CODES.P || !supportsDistortionCorrection()) {
+      return cameraManager
+          .getCameraCharacteristics(cameraDevice.getId())
+          .get(CameraCharacteristics.SENSOR_INFO_PIXEL_ARRAY_SIZE);
+    }
+    // Get the current distortion correction mode
+    Integer distortionCorrectionMode =
+        captureRequestBuilder.get(CaptureRequest.DISTORTION_CORRECTION_MODE);
+    // Return the correct boundaries depending on the mode
+    android.graphics.Rect rect;
+    if (distortionCorrectionMode == null
+        || distortionCorrectionMode == CaptureRequest.DISTORTION_CORRECTION_MODE_OFF) {
+      rect =
+          cameraManager
+              .getCameraCharacteristics(cameraDevice.getId())
+              .get(CameraCharacteristics.SENSOR_INFO_PRE_CORRECTION_ACTIVE_ARRAY_SIZE);
+    } else {
+      rect =
+          cameraManager
+              .getCameraCharacteristics(cameraDevice.getId())
+              .get(CameraCharacteristics.SENSOR_INFO_ACTIVE_ARRAY_SIZE);
+    }
+    return rect == null ? null : new Size(rect.width(), rect.height());
+  }
+
+  private boolean isExposurePointSupported() throws CameraAccessException {
+    Integer supportedRegions =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.CONTROL_MAX_REGIONS_AE);
+    return supportedRegions != null && supportedRegions > 0;
+  }
+
+  private boolean isFocusPointSupported() throws CameraAccessException {
+    Integer supportedRegions =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.CONTROL_MAX_REGIONS_AF);
+    return supportedRegions != null && supportedRegions > 0;
+  }
+
+  public double getMinExposureOffset() throws CameraAccessException {
+    Range<Integer> range =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.CONTROL_AE_COMPENSATION_RANGE);
+    double minStepped = range == null ? 0 : range.getLower();
+    double stepSize = getExposureOffsetStepSize();
+    return minStepped * stepSize;
+  }
+
+  public double getMaxExposureOffset() throws CameraAccessException {
+    Range<Integer> range =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.CONTROL_AE_COMPENSATION_RANGE);
+    double maxStepped = range == null ? 0 : range.getUpper();
+    double stepSize = getExposureOffsetStepSize();
+    return maxStepped * stepSize;
+  }
+
+  public double getExposureOffsetStepSize() throws CameraAccessException {
+    Rational stepSize =
+        cameraManager
+            .getCameraCharacteristics(cameraDevice.getId())
+            .get(CameraCharacteristics.CONTROL_AE_COMPENSATION_STEP);
+    return stepSize == null ? 0.0 : stepSize.doubleValue();
+  }
+
+  public void setExposureOffset(@NonNull final Result result, double offset)
+      throws CameraAccessException {
+    // Set the exposure offset
+    double stepSize = getExposureOffsetStepSize();
+    exposureOffset = (int) (offset / stepSize);
+    // Apply it
+    updateExposure(exposureMode);
+    this.cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
+    result.success(offset);
+  }
+
+  public float getMaxZoomLevel() {
+    return cameraZoom.maxZoom;
+  }
+
+  public float getMinZoomLevel() {
+    return CameraZoom.DEFAULT_ZOOM_FACTOR;
+  }
+
+  public void setZoomLevel(@NonNull final Result result, float zoom) throws CameraAccessException {
+    float maxZoom = cameraZoom.maxZoom;
+    float minZoom = CameraZoom.DEFAULT_ZOOM_FACTOR;
+
+    if (zoom > maxZoom || zoom < minZoom) {
+      String errorMessage =
+          String.format(
+              Locale.ENGLISH,
+              "Zoom level out of bounds (zoom level should be between %f and %f).",
+              minZoom,
+              maxZoom);
+      result.error("ZOOM_ERROR", errorMessage, null);
+      return;
+    }
+
+    //Zoom area is calculated relative to sensor area (activeRect)
+    if (captureRequestBuilder != null) {
+      final Rect computedZoom = cameraZoom.computeZoom(zoom);
+      captureRequestBuilder.set(CaptureRequest.SCALER_CROP_REGION, computedZoom);
+      cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
+    }
+
+    result.success(null);
+  }
+
+  public void lockCaptureOrientation(PlatformChannel.DeviceOrientation orientation) {
+    this.lockedCaptureOrientation = orientation;
+  }
+
+  public void unlockCaptureOrientation() {
+    this.lockedCaptureOrientation = null;
+  }
+
+  private void updateFpsRange() {
+    if (fpsRange == null) {
+      return;
+    }
+
+    captureRequestBuilder.set(CaptureRequest.CONTROL_AE_TARGET_FPS_RANGE, fpsRange);
+  }
+
+  private void updateFocus(FocusMode mode) {
+    if (useAutoFocus) {
+      int[] modes = cameraCharacteristics.get(CameraCharacteristics.CONTROL_AF_AVAILABLE_MODES);
+      // Auto focus is not supported
+      if (modes == null
+          || modes.length == 0
+          || (modes.length == 1 && modes[0] == CameraCharacteristics.CONTROL_AF_MODE_OFF)) {
+        useAutoFocus = false;
+        captureRequestBuilder.set(
+            CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_OFF);
+      } else {
+        // Applying auto focus
+        switch (mode) {
+          case locked:
+            captureRequestBuilder.set(
+                CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_AUTO);
+            break;
+          case auto:
+            captureRequestBuilder.set(
+                CaptureRequest.CONTROL_AF_MODE,
+                recordingVideo
+                    ? CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_VIDEO
+                    : CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
+          default:
+            break;
+        }
+        MeteringRectangle afRect = cameraRegions.getAFMeteringRectangle();
+        captureRequestBuilder.set(
+            CaptureRequest.CONTROL_AF_REGIONS,
+            afRect == null ? null : new MeteringRectangle[] {afRect});
+      }
+    } else {
+      captureRequestBuilder.set(CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_OFF);
+    }
+  }
+
+  private void updateExposure(ExposureMode mode) {
+    exposureMode = mode;
+
+    // Applying auto exposure
+    MeteringRectangle aeRect = cameraRegions.getAEMeteringRectangle();
+    captureRequestBuilder.set(
+        CaptureRequest.CONTROL_AE_REGIONS,
+        aeRect == null ? null : new MeteringRectangle[] {cameraRegions.getAEMeteringRectangle()});
+
+    switch (mode) {
+      case locked:
+        captureRequestBuilder.set(CaptureRequest.CONTROL_AE_LOCK, true);
+        break;
+      case auto:
+      default:
+        captureRequestBuilder.set(CaptureRequest.CONTROL_AE_LOCK, false);
+        break;
+    }
+
+    captureRequestBuilder.set(CaptureRequest.CONTROL_AE_EXPOSURE_COMPENSATION, exposureOffset);
   }
 
   private void updateFlash(FlashMode mode) {
     // Get flash
     flashMode = mode;
-    initPreviewCaptureBuilder();
-    try {
-      cameraCaptureSession.setRepeatingRequest(
-          captureRequestBuilder.build(), pictureCaptureCallback, null);
-    } catch (CameraAccessException e) {
-      pictureCaptureRequest.error("cameraAccess", e.getMessage(), null);
-    }
-  }
 
-  private void initPreviewCaptureBuilder() {
-    captureRequestBuilder.set(CaptureRequest.CONTROL_MODE, CaptureRequest.CONTROL_MODE_AUTO);
+    // Applying flash modes
     switch (flashMode) {
       case off:
         captureRequestBuilder.set(
@@ -701,8 +1076,6 @@ public class Camera {
         captureRequestBuilder.set(CaptureRequest.FLASH_MODE, CaptureRequest.FLASH_MODE_TORCH);
         break;
     }
-    captureRequestBuilder.set(
-        CaptureRequest.CONTROL_AF_MODE, CaptureRequest.CONTROL_AF_MODE_CONTINUOUS_PICTURE);
   }
 
   public void startPreview() throws CameraAccessException {
@@ -762,37 +1135,11 @@ public class Camera {
         null);
   }
 
-  public float getMaxZoomLevel() {
-    return cameraZoom.maxZoom;
-  }
-
-  public float getMinZoomLevel() {
-    return CameraZoom.DEFAULT_ZOOM_FACTOR;
-  }
-
-  public void setZoomLevel(@NonNull final Result result, float zoom) throws CameraAccessException {
-    float maxZoom = cameraZoom.maxZoom;
-    float minZoom = CameraZoom.DEFAULT_ZOOM_FACTOR;
-
-    if (zoom > maxZoom || zoom < minZoom) {
-      String errorMessage =
-          String.format(
-              Locale.ENGLISH,
-              "Zoom level out of bounds (zoom level should be between %f and %f).",
-              minZoom,
-              maxZoom);
-      result.error("ZOOM_ERROR", errorMessage, null);
-      return;
+  public void stopImageStream() throws CameraAccessException {
+    if (imageStreamReader != null) {
+      imageStreamReader.setOnImageAvailableListener(null, null);
     }
-
-    //Zoom area is calculated relative to sensor area (activeRect)
-    if (captureRequestBuilder != null) {
-      final Rect computedZoom = cameraZoom.computeZoom(zoom);
-      captureRequestBuilder.set(CaptureRequest.SCALER_CROP_REGION, computedZoom);
-      cameraCaptureSession.setRepeatingRequest(captureRequestBuilder.build(), null, null);
-    }
-
-    result.success(null);
+    startPreview();
   }
 
   private void closeCaptureSession() {
@@ -827,14 +1174,6 @@ public class Camera {
   public void dispose() {
     close();
     flutterTexture.release();
-    orientationEventListener.disable();
-  }
-
-  private int getMediaOrientation() {
-    final int sensorOrientationOffset =
-        (currentOrientation == ORIENTATION_UNKNOWN)
-            ? 0
-            : (isFrontFacing) ? -currentOrientation : currentOrientation;
-    return (sensorOrientationOffset + sensorOrientation + 360) % 360;
+    deviceOrientationListener.stop();
   }
 }
