@@ -1,12 +1,19 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.camera;
 
 import static junit.framework.TestCase.assertNull;
 import static org.junit.Assert.assertEquals;
 
 import androidx.annotation.NonNull;
+import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.StandardMethodCodec;
+import io.flutter.plugins.camera.types.ExposureMode;
+import io.flutter.plugins.camera.types.FocusMode;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.List;
@@ -58,7 +65,7 @@ public class DartMessengerTest {
 
   @Test
   public void sendCameraInitializedEvent_includesPreviewSize() {
-    dartMessenger.sendCameraInitializedEvent(0, 0);
+    dartMessenger.sendCameraInitializedEvent(0, 0, ExposureMode.auto, FocusMode.auto, true, true);
 
     List<ByteBuffer> sentMessages = fakeBinaryMessenger.getMessages();
     assertEquals(1, sentMessages.size());
@@ -66,6 +73,10 @@ public class DartMessengerTest {
     assertEquals("initialized", call.method);
     assertEquals(0, (double) call.argument("previewWidth"), 0);
     assertEquals(0, (double) call.argument("previewHeight"), 0);
+    assertEquals("ExposureMode auto", call.argument("exposureMode"), "auto");
+    assertEquals("FocusMode continuous", call.argument("focusMode"), "auto");
+    assertEquals("exposurePointSupported", call.argument("exposurePointSupported"), true);
+    assertEquals("focusPointSupported", call.argument("focusPointSupported"), true);
   }
 
   @Test
@@ -77,6 +88,17 @@ public class DartMessengerTest {
     MethodCall call = decodeSentMessage(sentMessages.get(0));
     assertEquals("camera_closing", call.method);
     assertNull(call.argument("description"));
+  }
+
+  @Test
+  public void sendDeviceOrientationChangedEvent() {
+    dartMessenger.sendDeviceOrientationChangeEvent(PlatformChannel.DeviceOrientation.PORTRAIT_UP);
+
+    List<ByteBuffer> sentMessages = fakeBinaryMessenger.getMessages();
+    assertEquals(1, sentMessages.size());
+    MethodCall call = decodeSentMessage(sentMessages.get(0));
+    assertEquals("orientation_changed", call.method);
+    assertEquals(call.argument("orientation"), "portraitUp");
   }
 
   private MethodCall decodeSentMessage(ByteBuffer sentMessage) {
