@@ -12,5 +12,8 @@ export 'package:camera_platform_interface/camera_platform_interface.dart'
         CameraException,
         CameraLensDirection,
         FlashMode,
+        ExposureMode,
+        FocusMode,
         ResolutionPreset,
-        XFile;
+        XFile,
+        ImageFormatGroup;
