@@ -3,10 +3,12 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:math';
 import 'dart:ui';
 
 import 'package:camera/camera.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -26,7 +28,18 @@ get mockAvailableCameras => [
 
 get mockInitializeCamera => 13;
 
-get mockOnCameraInitializedEvent => CameraInitializedEvent(13, 75, 75);
+get mockOnCameraInitializedEvent => CameraInitializedEvent(
+      13,
+      75,
+      75,
+      ExposureMode.auto,
+      true,
+      FocusMode.auto,
+      true,
+    );
+
+get mockOnDeviceOrientationChangedEvent =>
+    DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
 
 get mockOnCameraClosingEvent => null;
 
@@ -162,6 +175,22 @@ void main() {
       mockPlatformException = false;
     });
 
+    test('initialize() sets imageFormat', () async {
+      debugDefaultTargetPlatformOverride = TargetPlatform.android;
+      CameraController cameraController = CameraController(
+        CameraDescription(
+            name: 'cam',
+            lensDirection: CameraLensDirection.back,
+            sensorOrientation: 90),
+        ResolutionPreset.max,
+        imageFormatGroup: ImageFormatGroup.yuv420,
+      );
+      await cameraController.initialize();
+      verify(CameraPlatform.instance
+              .initializeCamera(13, imageFormatGroup: ImageFormatGroup.yuv420))
+          .called(1);
+    });
+
     test('prepareForVideoRecording() calls $CameraPlatform ', () async {
       CameraController cameraController = CameraController(
           CameraDescription(
@@ -184,12 +213,21 @@ void main() {
               sensorOrientation: 90),
           ResolutionPreset.max);
       expect(
-          cameraController.takePicture(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController.',
-            'takePicture was called on uninitialized CameraController',
-          )));
+        cameraController.takePicture(),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'takePicture() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('takePicture() throws $CameraException when takePicture is true',
@@ -257,12 +295,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          cameraController.startVideoRecording(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'startVideoRecording was called on uninitialized CameraController',
-          )));
+        cameraController.startVideoRecording(),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'startVideoRecording() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
     test('startVideoRecording() throws $CameraException when recording videos',
         () async {
@@ -321,12 +368,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          cameraController.getMaxZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMaxZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMaxZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMaxZoomLevel() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('getMaxZoomLevel() throws $CameraException when disposed', () async {
@@ -341,12 +397,21 @@ void main() {
       await cameraController.dispose();
 
       expect(
-          cameraController.getMaxZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMaxZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMaxZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Disposed CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMaxZoomLevel() was called on a disposed CameraController.',
+              ),
+        ),
+      );
     });
 
     test(
@@ -403,12 +468,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          cameraController.getMinZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMinZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMinZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMinZoomLevel() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('getMinZoomLevel() throws $CameraException when disposed', () async {
@@ -423,12 +497,21 @@ void main() {
       await cameraController.dispose();
 
       expect(
-          cameraController.getMinZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMinZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMinZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Disposed CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMinZoomLevel() was called on a disposed CameraController.',
+              ),
+        ),
+      );
     });
 
     test(
@@ -484,12 +567,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          () => cameraController.setZoomLevel(42.0),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'setZoomLevel was called on uninitialized CameraController',
-          )));
+        () => cameraController.setZoomLevel(42.0),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'setZoomLevel() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('setZoomLevel() throws $CameraException when disposed', () async {
@@ -504,12 +596,21 @@ void main() {
       await cameraController.dispose();
 
       expect(
-          () => cameraController.setZoomLevel(42.0),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'setZoomLevel was called on uninitialized CameraController',
-          )));
+        () => cameraController.setZoomLevel(42.0),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Disposed CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'setZoomLevel() was called on a disposed CameraController.',
+              ),
+        ),
+      );
     });
 
     test(
@@ -603,12 +704,508 @@ void main() {
             'This is a test error message',
           )));
     });
+
+    test('setExposureMode() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.setExposureMode(ExposureMode.auto);
+
+      verify(CameraPlatform.instance
+              .setExposureMode(cameraController.cameraId, ExposureMode.auto))
+          .called(1);
+    });
+
+    test('setExposureMode() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      when(CameraPlatform.instance
+              .setExposureMode(cameraController.cameraId, ExposureMode.auto))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.setExposureMode(ExposureMode.auto),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('setExposurePoint() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.setExposurePoint(Offset(0.5, 0.5));
+
+      verify(CameraPlatform.instance.setExposurePoint(
+              cameraController.cameraId, Point<double>(0.5, 0.5)))
+          .called(1);
+    });
+
+    test('setExposurePoint() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      when(CameraPlatform.instance.setExposurePoint(
+              cameraController.cameraId, Point<double>(0.5, 0.5)))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.setExposurePoint(Offset(0.5, 0.5)),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('getMinExposureOffset() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.getMinExposureOffset();
+
+      verify(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .called(1);
+    });
+
+    test('getMinExposureOffset() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      when(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.getMinExposureOffset(),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('getMaxExposureOffset() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.getMaxExposureOffset();
+
+      verify(CameraPlatform.instance
+              .getMaxExposureOffset(cameraController.cameraId))
+          .called(1);
+    });
+
+    test('getMaxExposureOffset() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      when(CameraPlatform.instance
+              .getMaxExposureOffset(cameraController.cameraId))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.getMaxExposureOffset(),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('getExposureOffsetStepSize() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.getExposureOffsetStepSize();
+
+      verify(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .called(1);
+    });
+
+    test(
+        'getExposureOffsetStepSize() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      when(CameraPlatform.instance
+              .getExposureOffsetStepSize(cameraController.cameraId))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.getExposureOffsetStepSize(),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('setExposureOffset() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => -1.0);
+      when(CameraPlatform.instance
+              .getMaxExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => 2.0);
+      when(CameraPlatform.instance
+              .getExposureOffsetStepSize(cameraController.cameraId))
+          .thenAnswer((_) async => 1.0);
+
+      await cameraController.setExposureOffset(1.0);
+
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 1.0))
+          .called(1);
+    });
+
+    test('setExposureOffset() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => -1.0);
+      when(CameraPlatform.instance
+              .getMaxExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => 2.0);
+      when(CameraPlatform.instance
+              .getExposureOffsetStepSize(cameraController.cameraId))
+          .thenAnswer((_) async => 1.0);
+      when(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 1.0))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.setExposureOffset(1.0),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test(
+        'setExposureOffset() throws $CameraException when offset is out of bounds',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => -1.0);
+      when(CameraPlatform.instance
+              .getMaxExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => 2.0);
+      when(CameraPlatform.instance
+              .getExposureOffsetStepSize(cameraController.cameraId))
+          .thenAnswer((_) async => 1.0);
+
+      expect(
+          cameraController.setExposureOffset(3.0),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'exposureOffsetOutOfBounds',
+            'The provided exposure offset was outside the supported range for this device.',
+          )));
+      expect(
+          cameraController.setExposureOffset(-2.0),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'exposureOffsetOutOfBounds',
+            'The provided exposure offset was outside the supported range for this device.',
+          )));
+
+      await cameraController.setExposureOffset(2.0);
+      await cameraController.setExposureOffset(-1.0);
+      await cameraController.setExposureOffset(-0.0);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 2.0))
+          .called(1);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, -1.0))
+          .called(1);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 0.0))
+          .called(1);
+    });
+
+    test('setExposureOffset() rounds offset to nearest step', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance
+              .getMinExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => -1.0);
+      when(CameraPlatform.instance
+              .getMaxExposureOffset(cameraController.cameraId))
+          .thenAnswer((_) async => 1.0);
+      when(CameraPlatform.instance
+              .getExposureOffsetStepSize(cameraController.cameraId))
+          .thenAnswer((_) async => 0.4);
+      when(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 1.0))
+          .thenAnswer((_) async => 1.0);
+
+      await cameraController.setExposureOffset(1.0);
+      await cameraController.setExposureOffset(-1.0);
+      await cameraController.setExposureOffset(0.1);
+      await cameraController.setExposureOffset(0.2);
+      await cameraController.setExposureOffset(0.3);
+      await cameraController.setExposureOffset(0.4);
+      await cameraController.setExposureOffset(0.5);
+      await cameraController.setExposureOffset(0.6);
+      await cameraController.setExposureOffset(0.7);
+      await cameraController.setExposureOffset(-0.1);
+      await cameraController.setExposureOffset(-0.2);
+      await cameraController.setExposureOffset(-0.3);
+      await cameraController.setExposureOffset(-0.4);
+      await cameraController.setExposureOffset(-0.5);
+      await cameraController.setExposureOffset(-0.6);
+      await cameraController.setExposureOffset(-0.7);
+
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 0.8))
+          .called(3);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, -0.8))
+          .called(3);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 0.0))
+          .called(2);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, 0.4))
+          .called(4);
+      verify(CameraPlatform.instance
+              .setExposureOffset(cameraController.cameraId, -0.4))
+          .called(4);
+    });
+
+    test('lockCaptureOrientation() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.lockCaptureOrientation();
+      expect(cameraController.value.lockedCaptureOrientation,
+          equals(DeviceOrientation.portraitUp));
+      await cameraController
+          .lockCaptureOrientation(DeviceOrientation.landscapeRight);
+      expect(cameraController.value.lockedCaptureOrientation,
+          equals(DeviceOrientation.landscapeRight));
+
+      verify(CameraPlatform.instance.lockCaptureOrientation(
+              cameraController.cameraId, DeviceOrientation.portraitUp))
+          .called(1);
+      verify(CameraPlatform.instance.lockCaptureOrientation(
+              cameraController.cameraId, DeviceOrientation.landscapeRight))
+          .called(1);
+    });
+
+    test(
+        'lockCaptureOrientation() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance.lockCaptureOrientation(
+              cameraController.cameraId, DeviceOrientation.portraitUp))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.lockCaptureOrientation(DeviceOrientation.portraitUp),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('unlockCaptureOrientation() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.unlockCaptureOrientation();
+      expect(cameraController.value.lockedCaptureOrientation, equals(null));
+
+      verify(CameraPlatform.instance
+              .unlockCaptureOrientation(cameraController.cameraId))
+          .called(1);
+    });
+
+    test(
+        'unlockCaptureOrientation() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance
+              .unlockCaptureOrientation(cameraController.cameraId))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.unlockCaptureOrientation(),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
   });
 }
 
 class MockCameraPlatform extends Mock
     with MockPlatformInterfaceMixin
     implements CameraPlatform {
+  @override
+  Future<void> initializeCamera(int cameraId,
+      {ImageFormatGroup imageFormatGroup});
+
   @override
   Future<List<CameraDescription>> availableCameras() =>
       Future.value(mockAvailableCameras);
@@ -636,12 +1233,17 @@ class MockCameraPlatform extends Mock
       Stream.value(mockOnCameraErrorEvent);
 
   @override
+  Stream<DeviceOrientationChangedEvent> onDeviceOrientationChanged() =>
+      Stream.value(mockOnDeviceOrientationChangedEvent);
+
+  @override
   Future<XFile> takePicture(int cameraId) => mockPlatformException
       ? throw PlatformException(code: 'foo', message: 'bar')
       : Future.value(mockTakePicture);
 
   @override
-  Future<XFile> startVideoRecording(int cameraId) =>
+  Future<XFile> startVideoRecording(int cameraId,
+          {Duration maxVideoDuration}) =>
       Future.value(mockVideoRecordingXFile);
 }
 
