@@ -7,6 +7,7 @@ import 'dart:ui';
 import 'package:camera/camera.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
@@ -21,6 +22,11 @@ void main() {
         isTakingPicture: false,
         isStreamingImages: false,
         flashMode: FlashMode.auto,
+        exposureMode: ExposureMode.auto,
+        exposurePointSupported: true,
+        deviceOrientation: DeviceOrientation.portraitUp,
+        lockedCaptureOrientation: DeviceOrientation.portraitUp,
+        recordingOrientation: DeviceOrientation.portraitUp,
       );
 
       expect(cameraValue, isA<CameraValue>());
@@ -31,6 +37,13 @@ void main() {
       expect(cameraValue.isRecordingVideo, isFalse);
       expect(cameraValue.isTakingPicture, isFalse);
       expect(cameraValue.isStreamingImages, isFalse);
+      expect(cameraValue.flashMode, FlashMode.auto);
+      expect(cameraValue.exposureMode, ExposureMode.auto);
+      expect(cameraValue.exposurePointSupported, true);
+      expect(cameraValue.deviceOrientation, DeviceOrientation.portraitUp);
+      expect(
+          cameraValue.lockedCaptureOrientation, DeviceOrientation.portraitUp);
+      expect(cameraValue.recordingOrientation, DeviceOrientation.portraitUp);
     });
 
     test('Can be created as uninitialized', () {
@@ -44,6 +57,12 @@ void main() {
       expect(cameraValue.isRecordingVideo, isFalse);
       expect(cameraValue.isTakingPicture, isFalse);
       expect(cameraValue.isStreamingImages, isFalse);
+      expect(cameraValue.flashMode, FlashMode.auto);
+      expect(cameraValue.exposureMode, null);
+      expect(cameraValue.exposurePointSupported, false);
+      expect(cameraValue.deviceOrientation, DeviceOrientation.portraitUp);
+      expect(cameraValue.lockedCaptureOrientation, null);
+      expect(cameraValue.recordingOrientation, null);
     });
 
     test('Can be copied with isInitialized', () {
@@ -59,6 +78,11 @@ void main() {
       expect(cameraValue.isTakingPicture, isFalse);
       expect(cameraValue.isStreamingImages, isFalse);
       expect(cameraValue.flashMode, FlashMode.auto);
+      expect(cameraValue.exposureMode, null);
+      expect(cameraValue.exposurePointSupported, false);
+      expect(cameraValue.deviceOrientation, DeviceOrientation.portraitUp);
+      expect(cameraValue.lockedCaptureOrientation, null);
+      expect(cameraValue.recordingOrientation, null);
     });
 
     test('Has aspectRatio after setting size', () {
@@ -66,7 +90,7 @@ void main() {
       var cameraValue =
           cv.copyWith(isInitialized: true, previewSize: Size(20, 10));
 
-      expect(cameraValue.aspectRatio, 0.5);
+      expect(cameraValue.aspectRatio, 2.0);
     });
 
     test('hasError is true after setting errorDescription', () {
@@ -97,10 +121,17 @@ void main() {
         isTakingPicture: false,
         isStreamingImages: false,
         flashMode: FlashMode.auto,
+        exposureMode: ExposureMode.auto,
+        focusMode: FocusMode.auto,
+        exposurePointSupported: true,
+        focusPointSupported: true,
+        deviceOrientation: DeviceOrientation.portraitUp,
+        lockedCaptureOrientation: DeviceOrientation.portraitUp,
+        recordingOrientation: DeviceOrientation.portraitUp,
       );
 
       expect(cameraValue.toString(),
-          'CameraValue(isRecordingVideo: false, isInitialized: false, errorDescription: null, previewSize: Size(10.0, 10.0), isStreamingImages: false, flashMode: FlashMode.auto)');
+          'CameraValue(isRecordingVideo: false, isInitialized: false, errorDescription: null, previewSize: Size(10.0, 10.0), isStreamingImages: false, flashMode: FlashMode.auto, exposureMode: ExposureMode.auto, focusMode: FocusMode.auto, exposurePointSupported: true, focusPointSupported: true, deviceOrientation: DeviceOrientation.portraitUp, lockedCaptureOrientation: DeviceOrientation.portraitUp, recordingOrientation: DeviceOrientation.portraitUp)');
     });
   });
 }
