@@ -234,82 +234,63 @@ void main() {
     });
 
     test('videoEventsFor', () async {
-      // TODO(cbenhagen): This has been deprecated and should be replaced
-      // with `ServicesBinding.instance.defaultBinaryMessenger` when it's
-      // available on all the versions of Flutter that we test.
-      // ignore: deprecated_member_use
-      defaultBinaryMessenger.setMockMessageHandler(
+      ServicesBinding.instance.defaultBinaryMessenger.setMockMessageHandler(
         "flutter.io/videoPlayer/videoEvents123",
         (ByteData message) async {
           final MethodCall methodCall =
               const StandardMethodCodec().decodeMethodCall(message);
           if (methodCall.method == 'listen') {
-            // TODO(cbenhagen): This has been deprecated and should be replaced
-            // with `ServicesBinding.instance.defaultBinaryMessenger` when it's
-            // available on all the versions of Flutter that we test.
-            // ignore: deprecated_member_use
-            await defaultBinaryMessenger.handlePlatformMessage(
-                "flutter.io/videoPlayer/videoEvents123",
-                const StandardMethodCodec()
-                    .encodeSuccessEnvelope(<String, dynamic>{
-                  'event': 'initialized',
-                  'duration': 98765,
-                  'width': 1920,
-                  'height': 1080,
-                }),
-                (ByteData data) {});
+            await ServicesBinding.instance.defaultBinaryMessenger
+                .handlePlatformMessage(
+                    "flutter.io/videoPlayer/videoEvents123",
+                    const StandardMethodCodec()
+                        .encodeSuccessEnvelope(<String, dynamic>{
+                      'event': 'initialized',
+                      'duration': 98765,
+                      'width': 1920,
+                      'height': 1080,
+                    }),
+                    (ByteData data) {});
 
-            // TODO(cbenhagen): This has been deprecated and should be replaced
-            // with `ServicesBinding.instance.defaultBinaryMessenger` when it's
-            // available on all the versions of Flutter that we test.
-            // ignore: deprecated_member_use
-            await defaultBinaryMessenger.handlePlatformMessage(
-                "flutter.io/videoPlayer/videoEvents123",
-                const StandardMethodCodec()
-                    .encodeSuccessEnvelope(<String, dynamic>{
-                  'event': 'completed',
-                }),
-                (ByteData data) {});
+            await ServicesBinding.instance.defaultBinaryMessenger
+                .handlePlatformMessage(
+                    "flutter.io/videoPlayer/videoEvents123",
+                    const StandardMethodCodec()
+                        .encodeSuccessEnvelope(<String, dynamic>{
+                      'event': 'completed',
+                    }),
+                    (ByteData data) {});
 
-            // TODO(cbenhagen): This has been deprecated and should be replaced
-            // with `ServicesBinding.instance.defaultBinaryMessenger` when it's
-            // available on all the versions of Flutter that we test.
-            // ignore: deprecated_member_use
-            await defaultBinaryMessenger.handlePlatformMessage(
-                "flutter.io/videoPlayer/videoEvents123",
-                const StandardMethodCodec()
-                    .encodeSuccessEnvelope(<String, dynamic>{
-                  'event': 'bufferingUpdate',
-                  'values': <List<dynamic>>[
-                    <int>[0, 1234],
-                    <int>[1235, 4000],
-                  ],
-                }),
-                (ByteData data) {});
+            await ServicesBinding.instance.defaultBinaryMessenger
+                .handlePlatformMessage(
+                    "flutter.io/videoPlayer/videoEvents123",
+                    const StandardMethodCodec()
+                        .encodeSuccessEnvelope(<String, dynamic>{
+                      'event': 'bufferingUpdate',
+                      'values': <List<dynamic>>[
+                        <int>[0, 1234],
+                        <int>[1235, 4000],
+                      ],
+                    }),
+                    (ByteData data) {});
 
-            // TODO(cbenhagen): This has been deprecated and should be replaced
-            // with `ServicesBinding.instance.defaultBinaryMessenger` when it's
-            // available on all the versions of Flutter that we test.
-            // ignore: deprecated_member_use
-            await defaultBinaryMessenger.handlePlatformMessage(
-                "flutter.io/videoPlayer/videoEvents123",
-                const StandardMethodCodec()
-                    .encodeSuccessEnvelope(<String, dynamic>{
-                  'event': 'bufferingStart',
-                }),
-                (ByteData data) {});
+            await ServicesBinding.instance.defaultBinaryMessenger
+                .handlePlatformMessage(
+                    "flutter.io/videoPlayer/videoEvents123",
+                    const StandardMethodCodec()
+                        .encodeSuccessEnvelope(<String, dynamic>{
+                      'event': 'bufferingStart',
+                    }),
+                    (ByteData data) {});
 
-            // TODO(cbenhagen): This has been deprecated and should be replaced
-            // with `ServicesBinding.instance.defaultBinaryMessenger` when it's
-            // available on all the versions of Flutter that we test.
-            // ignore: deprecated_member_use
-            await defaultBinaryMessenger.handlePlatformMessage(
-                "flutter.io/videoPlayer/videoEvents123",
-                const StandardMethodCodec()
-                    .encodeSuccessEnvelope(<String, dynamic>{
-                  'event': 'bufferingEnd',
-                }),
-                (ByteData data) {});
+            await ServicesBinding.instance.defaultBinaryMessenger
+                .handlePlatformMessage(
+                    "flutter.io/videoPlayer/videoEvents123",
+                    const StandardMethodCodec()
+                        .encodeSuccessEnvelope(<String, dynamic>{
+                      'event': 'bufferingEnd',
+                    }),
+                    (ByteData data) {});
 
             return const StandardMethodCodec().encodeSuccessEnvelope(null);
           } else if (methodCall.method == 'cancel') {
